Add biquad service dunctions

+++ basic_float_math.h
@@ -22,7 +22,7 @@ INLINE void Interleave(float_t* in0, float_t* in1, float2_t* out, intg_t numSamp
 
 INLINE void SmoothValue(float_t* current, float_t* target, float_t step) {
     float_t delta = *target - *current;
-    if (fabs(delta) <= step) {
+    if (fabsf(delta) <= step) {
         *current = *target;
     }
     else {
